calypso_AgentiArena/bid-engine/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
requests
